Create different unit test for the controller methods

+++ AddressBookAPI/Controllers/ContactsController.cs
@@ -20,32 +20,49 @@ namespace AddressBookAPI.Controllers
         }
 
         [HttpGet]
-        public List<ContactDTO> Get()
+        public ActionResult<List<ContactDTO>> Get()
         {
             var contacts = _repository.GetAll();
-            return mapper.Map<List<ContactDTO>>(contacts);
+            return Ok(mapper.Map<List<ContactDTO>>(contacts));
         }
 
         [HttpGet("{id:int}", Name = "getContact")]
-        public ContactDTO Get(int id)
+        public ActionResult<ContactDTO> Get(int id)
         {
             var contact = _repository.GetById(id);
-            return mapper.Map<ContactDTO>(contact);
+
+            if(contact == null)
+            {
+                return NotFound();
+            }
+            return Ok(mapper.Map<ContactDTO>(contact));
         }
 
         [HttpPost]
         public  ActionResult Post(ContactCreationDTO contactCreationDTO)
         {
-            var contact = mapper.Map<Contact>(contactCreationDTO);
-             _repository.Insert(contact);
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
 
-            return NoContent();
+            var contact = mapper.Map<Contact>(contactCreationDTO);
+            var newContact = mapper.Map<ContactDTO>(_repository.Insert(contact));
+
+            return CreatedAtAction(nameof(Get), new {id = newContact.Id}, newContact);
         }
 
 
         [HttpPut("{id:int}")]
         public ActionResult Put(ContactCreationDTO contactCreationDTO, int id)
         {
+            var existingContact = _repository.GetById(id);
+
+            if (existingContact == null)
+            {
+                return NotFound();
+            }
+
             var contact = mapper.Map<Contact>(contactCreationDTO);
             _repository.Update(contact, id);
 
@@ -55,6 +72,13 @@ namespace AddressBookAPI.Controllers
         [HttpDelete("{id:int}")]
         public ActionResult Delete(int id)
         {
+            var existingContact = _repository.GetById(id);
+
+            if(existingContact == null)
+            {
+                return NotFound();
+            }
+
             _repository.Delete(id);
             return NoContent();
         }
